Adding Recipe Entities
The RecipeEntity represents a recipe as stored in out system and present through our API. We also need a RecipesResponseEntity to represent how the collection of recipes are delivered to our application.

You could argue that the the response objects could live in the domain module - however because they represent the state of the whole system outside of our specific application I think it more fitting that they remain in the entities module; they represent an entity outside of our domain.

The SRP states we should group those things likely to change together. APIs are generally fairly stable and thus this response can live in the inner most circle of our application being the most stable and independent. It would also seem likely that were the response to change then the recipe entity may also change with it. I also chose not to map and of the other response properties found in the JSON such as title, version and href as they are not needed at this point in the application and decouples us from changes in this part of schema.