Adding ReactiveList
I am aiming to build a reactive architecture using RxJava as a major component. One feature of Rx is that an error will terminate a stream. Often this is the desired response, but there are many cases in mobile apps where as a subscriber you want to receive the error and not have to initialise a new stream to request more data.

One example would be displaying a list of products. If on the second page you encountered a network error you simply want to display a message to user and allow them to retry loading the next page. Rather than having to resubscribe -  and in the case of a stream that accumulated page items as each one was loaded, recreate your state - you can simply be provided the previous value and the error and then allow the user to trigger another load without any of the overhead of setting up your streams and subscriptions.

Because we are trying to seperate our UI layer from our entities, a lot of business logic involves mapping lists of entities into presentation objects more suitable for the UI to consume. And because we want our data objects to be immutable to avoid state based bugs, we need a way to iterate over a reactive list and convert it to a reactive list of another kind. We can do this using Kotlin's map function but doing it in the client code creates a lot of code reuse and local state with the storing of mapped list items and creation of new ReactiveLists with these mapped item. This can easily be encapsulated in our ReactiveLists class as done here with a custom map implementation

Often presentation logic will need to branch based on whether our ReactiveList has an error or not, or has a value or not. These convenience methods will help to reduce null checks and make the client code read more declaratively.

There will also be times when we want to process a list more than a simple mapping operation but still ultimately map the result to a ReactiveList of a different datatype. Here we can use the process method will be provide the current value to a function that will return the processed list - encapsulating the recreation of the new reactivelist