Adding ReactiveOperators extension function
One function of our ReactiveList is to capture errors in a stream into our data type without completing the stream. We can achieve this in a declarative manner using an Extension function on rx.Observable utilising an rx Transformer - a method for essentially writing our own operators. In this initial case we simply switchMap our input stream to an observable provided by the function param to our extension method. Here we can capture any error before it completes the parent stream. We also have an opportunity to store the previous value, and then when an error occurs combine the previous value with the error before passing it down stream.

This pattern is particularly useful for making network calls. Network calls are naturally streams that likely to error, but also likely to be initialised by a parent stream representing a load / refresh action.