Adding RecipesRepository
Our getRecipes component will need some way to retrieve our recipes. We have already defined our service interface but a repository pattern provides additional flexibility. It allows us to encapsulate our business rules around retrieval and storage. In our initial implementation we simply request the recipes from a service and pass them to any observer. In the future we may wish to implement a cache which we fall back to on network error.

It also provides us a way to handle our error handling an retry logic.  In the initial implementation we have utilised our switchMapReactiveList extension method to capture errors into our ReactiveList data type, and also exposed a refresh method to trigger a request to the service.

One thing to note about the initial implementation. We do not subscribe to the recipes observable in the repository. This makes the observable clod, meaning that calls to refresh will not trigger a service call unless there is a subscriber to the recipes observable. It also means that should there be be multiple subscriber to the stream, a second subscriber would need to make another call to refresh in order receive updates.

However, because of our initial use case for the repository being that there should only be one subscriber this seems like an appropriate implementation. The benefit of the repository pattern is that should these requirements change we can implement the changes in one place.