Adding Recipe Data Class
This class will represent a recipe to the UI. One of the principles of clean architecture is that the business layer should return data in the format most useful for the client. In this case that data is almost identical to the entity. The only change is that we present the ingredients as a list rather than a comma separated string. This stops our client needing to do unnecessary processing on the data in order to display it leaving our views to deal only with the presentation of the data, not the manipulation of it. It also makes our view layer much easier to test. In this case the example is trivial but for other cases it will not be as straight forward which is why implementing this patten is important.

The data structure is a kotlin data class, which is useful because of the equals and hashcode generation -this is particularly useful in the view layer for things such as diff utils. All the properties are final making the class immutable. Immutable data leads to less state bugs. It also forces the manipulation of state back into the business layer which helps us to organise our architecture as a reactive one. Our view layer should only be concerned with presenting data and providing use interaction back to the business layer so that it can manipulate the state of the application