Adding GetRecipesCommand
This class is the main class that holds our business rules. Here we request our recipes from the repository before ordering them by whether they have an image, then mapping them into our Recipes Data structure. The class itself is internal, meaning that it cannot be instantiated outside of our package. This helps with component cohesion and forces only classes that are designed to collaborate to be able to collaborate - in this case our GetRecipes class and the repository.

This class is also the class that will be used by clients of the getRecipes component, but because it is an internal class it will only be exposed by its parent type, a kotlin function (()->Observable<ReactiveList<Recipe>>). The advantage of this is that clients can be injected with any implementation of the function, allowing us to use polymorphism and decoupling the client from our implementation while also stopping others from using our concrete implementation directly and coupling various components in the application.

One other thing worth notting is the use of kotlin as Sequence when mapping and sorting our entities and splitting our ingredients. Because we are processing the lists in multiple ways using asSequence reduces the time complexity from O(a * n) where a is the number of operations performed on the list (filter, map etc)  to O(n), ie we only iterate over the lists once, vs once for each operation.