Adding LoadRecipesCommand
We need a way to request that we load / reload our recipes. We already have the functionality implemented in our repository but we do not want to expose our repository to clients of our getRecipes component. There fore we can use the same pattern as out GetRecipes class and write an internal class that implements kotlin function to (initially) simply proxy a call to our repository.

The class will be exposed to clients though it kotlin function : ()->Unit