Adding JavaX annotation processor
In order to have our classes injectable we need annotate them with @Inject. We could write our own provider methods to do this in order to remove the dependency from the module and there is an argument to be made for a class not knowing how it is injected but the convenience of simply annotating a constructor is a fair compromise in this situation as it does not actually effect the compiled code of the class it annotates.