Adding Dependency Injection Module for getRecipes Component
In order to inject our getRecipes component into various parts of the app we will need to define a module and provider methods. The goal is to provide our commands via their kotlin function signature, and thus encapsulate their implementation from the rest of application. Because we a re using a common signature we will use Qualifier classes to avoid any potential clashes.

We also need to define a scope for our repository. If we did not do this we would provide different instances to the load and get commands meaning they would not work correctly. On the other hand because of the nature of our commands and them not holding state this would be unnecessary on top of the fact we should in the initial implementation only require one instance of each.

The module was defined inside the domain.getRecipes package in order appease kotlin's internal class restrictions, however once the class was changed to java these restrictions did not apply and it could be moved back into the app module with the other Di classes. However, I decide to keep it here for 2 reasons. Firstly it makes sense to keep the dependency injection with the classes it injects, to improve component cohesion Secondly when a project scale it may be possible to generate the dagger classes for each module and save compilation time.

Finally the class is written in Java rather than kotlin. This is because dagger seemed to have a problem with providing the commands as functions. When I tried to write the module in kotlin and set the return types of the getRecipes commands to  ()->Observable<ReactiveList<Recipe>> the dagger compiler complained that there was not a provider for Function0<? extends Observable<ReactiveList<Recipe>>. However, when written in Java the module works. This is down to the differing ways that kotlin and java handle covariance and contravariance.