Adding ServicesModule to provide Recipes Service
We will use our ServiceFactory to create an instance of our retrofit service. However, in order to decouple our domain from our retrofit implementation of service calls our domain has defined its own Service interface. Because retrofit requires the annotations to work and does not allow inheritance, we can still provide service as defined in our domain component by creating a seperate retrofit service and providing the domain implementation via a delegation strategy.

This has the downside of creating an extra object and overhead / boiler plate for service creation but the trade off is worth it in order to maintain the decoupling of our domain from the detail of network layer implementation