Adding RecipesViewModel
Adding our recipes view model. This will use our defined getRecipes and loadRecipes commands to load the recipes and present them to the view. We also provide a refresh method to bind to any refresh or error buttons allowing the user to reload recipes.

It is important to not we clear subscriptions in onCleared. Failure to do this can cause a memory leak