Adding Resources to SystemServicesModule
Because our view model requires android resources in order to get the error message we must provide it here. Because the root of our graph is the application dagger can provide it to our provides method and we can then extract the resources.

you should use caution when using the application context to get resources, but because resources is in fact a global singleton using the app context is appropriate here