Adding ViewModelsModule
This view models module will be used to inject our view model. Because we are using android architecture view models as our base classes we actually need to provide a factory and allow our activities to request the instance from the factory. This allows the architecture components to manager the lifecycle of the view model inline with the activity it is being used in