Adding BaseActivity
Base Activity will serve as a common location to handle common activity features. In particular we can handle our injection into the AndroidDagger graph meaning all that is required for an activity to be injected is that it simple extends BaseActivity. It is important to not that the static call to inject must happen before super.onCreate