Adding DataBinding to the build
Databinding allows us to have a clearer separation of concerns between our view model and our view. Effectively it allows us to create a presentation layer that is more reusable. Without databinding we would need to write write handlers for updates on our view model the specify how that data is presented in the view. Using databinding this can be more clearly and succinctly defined inside out xml layouts