Adding FadeVisible Binding
One common presentation layer operation is to hide an show a view based on some kind of state change. While we could simply use ad data binding conditional to set a views visibility prorety in line we can achieve a nicer effect using animations which is not possible inline. By writing our own custom binding adapter called fadeVisible we can manage the animation and visibility setting in one location and reuse it any layout going forward.

One of the chaellenges of this approach is where to store references to the animation so that when the state changes we can simply cancel the first animation. In the past I have done this using a views tags which has the benefit of linking it directly with the view which is provided to out binding function. The down side of this is that multiple bindings use the approach they can not be applied together. This time I chose to use a weak hash map to store the animation. We need to make sure we use a weak hash map so that the map does not leak the view