Binding Recipes View Model  to view in Activity
I decided to move the recipes activity into its own package to help with component cohesion between it and the view model and other views that will interact directly with it. We inject our view model factory and inflate our layout using the binding utils. It is important to set the lifecycle owner of the bindings for the view model to work with the data bindings