Adding Loading Status Data Class
A common need for our view will be representing state. As alot of data will need to be loaded wither over the network or from storage this may take some time and we will need to communicate this to the user. There is also the possibility that the operation will fail so we need to represent that in our data structure too.

Kotlin sealed classes are the perfect mechanism for this. They Behave like a cross between a normal class and an enum class giving you compile time safety when iterating over the parent type