Adding AspectRatioImageView
Often you will need an image view that rather than scaling to the size of the provided image will scale to either its image or content but still maintain a give aspect ratio. This class aims to solve that issue. You can simply provide the ratio, 1 being square and the direction you with the image to scale.