Adding Picasso
Picasso library provides an easy way to load images and cache them. To handle this with databinding we can provide a simple binding adapter that initialises a Picasso instance and loads the provided url into the provided image. We can also provide a placeholder color to display while the image is loading.