Adding Recipe With Image View
This view will be used to display those recipes that have an image. We will use the android card view to provide a tile like layout. Because not all of the images are the same size or ratio ive chosen to use an aspect ratio of 1. This was the most common from the images I saw so seemed like the easiest choice.

I also chose to add a backing gradient to the text to make it more legible