Adding RecipesDiffUtil
to get maximum performance from our list of recipes and also if we want to add any animations we need to crate a custom diff util callback. In our case the contents of our recipes will not change so we can simply implement the areItemsTheSame and areContentTheSame methods in the same way by comparing the items directly for equality. Because they are kotlin data classes items with the same content will be equal.