Adding RecyclerView Spacing Decoration
While we could have handled spacing inside the views them selves it become inflexible quickly. That is why it is best practice to use a Decoration to achieve this padding. Here we simply type check the given item (as we only want to apply spacing to our recipes with images). We then apply different areas of padding depending on whether the item is in the first last or middle column to give an even padding around all our recipe tiles