Adding Recipes Adapter
A simple adapter implementation using our predefined view holder. We simply check a recipes image for emptiness to determine the view holder and delegate the binding to that class. We pass in the Binding instance as this couple our view holder to a specific view, which we want. Otherwise if we were to pass in an inflated view then create the binding inside the viewholder we could have class cast exceptions