Adding RecipesRecycler
This will be our custom view that pulls together our adapter and decorations. We use a lazy implementation took up the columns from resources. This allows us only make the method call once and save the value but using the resources allows us to specify different values for landscape and all other configuration. We could have used a portrait version but this would mean that had a device been square we would not have had a value.

It is also here that we define our spanSizeLookup. This is vital for our desired layout. Here we specify that all Text only recipes should span all columns, while anything else (image tiles) should span only the one.