Adding RecipeSelected callback to view model
We need a way for our view to open our recipe. We could always directly launch it from the adapter or view holder. The problem with this is often you will need to perform so other action, tracking for example. Its also breaks the single responsibility principle and forces our view to do anything other than present data. Because of this we want the callback to make its way to the view model where we can dermine the appropriate action.

In this case we will eventually want to launch a new screen. This requires using the Android framework and we should strive to reduce our dependency on the framework in our viewmodel, in order to make it easier to test. The way I've chosen to achieve this is to defined the callback on the view model, which is already being bound to our view and allow an API for our framework controller (activity / fragment) to provide a function to actually handle the navigation.

This moves the helps seperate the concerns of the view model and the framework controller