Binding RecipeSelected back through to viewmodel
Passing the function callback through view bindings and recycler/adapter to the views. The One thing to note is that setting the callback on the recycler does not invoke a dataset changed. This may lead to issues if the callback changes and is rebound or is bound after views have already been inflated. However, the benefit is that we do not have any unnecessary re-renders and because this value is a final value on the view model that should not be rebound we can have a simple setter. If the requirements change and this callback can change or needs to be bound later than the initial dataset then you will need to call dataSet changed to recreate the viewholder