Adding Recipe Headers
In order to show headers we need to add another decorator to our recycler view. We could have created arbitrary data types and passed them to the adapter or eve set boolean and handled the logic inside the adapter but this would be complicated and very coupled to the adapter, not to mention beyond its scope. Instead we can simply create a decoration that can easily be added and removed.

The decoration works by firstly creating header views for each type of header, the when called iterating over all children to find those that are the first of their type. From there we can work out the appropriate header and render it.

One advantage of this implementation is that it does not rely on any ordering logic, ie that the list has been sorted with images first. However, because we need to get the previous type when we pass a position -1 to our recycler adapter we run the risk of IndexOutOfBounds exception. To combat this I've updated the adapter to check for IndexOutOfBounds exceptions and return Type Unknown in this case. It also relies on that specific adapter being set and casting accordingly however it is an internal class and actually handles the fallback to unknown gracefully by defaulting to TYPE_UNKOWN and thus would simply not show headers should we change the adapter type