Adding RecipeDetailsActivity
Adding Activity to display the recipe details. I've chosen to use a Launcher Pattern here. This helps seperate our data serialisation needed to start an activity and helps keep the activity decluttered. Its also provided a more fluent API for clients to  launch the activity rather than having to manually handle adding and removing the extras from the intent.

We could have used made recipes serialisable, but there are costs to incur with this and keeping the version up to date can result in unexpected bugs. Parcelable is not possible if we wish to decouple our domain logic from the Android API so this appears to be the best solution