Adding IngredientsList
For each recipe we need to list out ingredients. There are multiple ways to achieve this. We could have used a recycler but having it nested inside a scrollview would have removed the benefit. Plus the size of the dataset does not really warrant a recycler. We could have used a binding on linear layout and handled the iteration and creations of loops in there but a custom view to encapsulate this seems more appropriate and more descriptive from the clients perspective.